- Sending the text-label file with labeled data - It was created the method labeled data at PrepData class
