+++ README.md
@@ -1,8 +1,7 @@
 # Native-Language-Classification
 The main proporse of this project is to develop an application that is able to identify and classify 26 Brazilian native languages. 
 
-The dataset consists of 27 samples of the New Testament from the Holy Bible. <br>
-The [indigenous](https://myswordbrasil.blogspot.com/2016/09/26-biblias-em-idiomas-indigenas-tupi-guarani.htm) and the [portuguese](https://myswordbrasil.blogspot.com/2017/06/biblias-para-mysword.html) version of the New Testament were found at My Sword Brasil.
+The dataset consists of 27 samples of the New Testament from the Holy Bible. The [indigenous](https://myswordbrasil.blogspot.com/2016/09/26-biblias-em-idiomas-indigenas-tupi-guarani.htm) and the [portuguese](https://myswordbrasil.blogspot.com/2017/06/biblias-para-mysword.html) version of the New Testament were found at My Sword Brasil.
 
 
 This is important to note that it was necessary to preprocess the verses of the bible removing some characters (i.e [] () <>) and align the
